game-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiomysql==0.2.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.0